Create the record class that will behave like a struct and has
